Moved the logic for checking to see if a timeblock was selected previously.
